No longer making db calls on main thread

+++ app/src/main/java/com/syftapp/codetest/data/repository/PostsRemoteMediator.kt
@@ -1,6 +1,5 @@
 package com.syftapp.codetest.data.repository
 
-import android.util.Log
 import androidx.paging.ExperimentalPagingApi
 import androidx.paging.LoadType
 import androidx.paging.PagingState
@@ -13,6 +12,7 @@ import com.syftapp.codetest.data.model.domain.Post
 import com.syftapp.codetest.data.model.domain.PostKey
 import io.reactivex.Single
 import io.reactivex.schedulers.Schedulers
+import java.io.InvalidObjectException
 
 @OptIn(ExperimentalPagingApi::class)
 class PostsRemoteMediator(
@@ -27,59 +27,56 @@ class PostsRemoteMediator(
         loadType: LoadType,
         state: PagingState<Int, Post>
     ): Single<MediatorResult> {
-        val page: Int = when (loadType) {
-            LoadType.REFRESH -> {
-                val postKey = getRemoteKeyClosestToCurrentPosition(state)
-                Log.e("LoadType", "LoadType is refresh")
+        return Single.just(loadType)
+            .subscribeOn(Schedulers.io())
+            .map {
+                when (it) {
+                    LoadType.REFRESH -> {
+                        val remoteKeys = getRemoteKeyClosestToCurrentPosition(state)
 
-                postKey?.nextKey?.minus(1) ?: -1
-            }
+                        remoteKeys?.nextKey?.minus(1) ?: 1
+                    }
+                    LoadType.PREPEND -> {
+                        val remoteKeys = getRemoteKeyForFirstItem(state)
+                            ?: throw InvalidObjectException("Result is empty")
 
-            LoadType.PREPEND -> {
-                val postKey = getRemoteKeyForFirstItem(state)
-                Log.e("LoadType", "LoadType is PREPEND")
+                        remoteKeys.previousKey ?: -1
+                    }
+                    LoadType.APPEND -> {
+                        val remoteKeys = getPostKeyForLastItem(state)
+                            ?: throw InvalidObjectException("Result is empty")
 
-                val prevKey = postKey?.previousKey ?: return Single.just(
-                    MediatorResult.Success(endOfPaginationReached = postKey != null)
-                )
-
-                prevKey
-            }
-
-            LoadType.APPEND -> {
-                val postKey = getPostKeyForLastItem(state)
-                Log.e("LoadType", "LoadType is APPEND")
-
-                val nextKey = postKey?.nextKey
-                    ?: return Single.just(MediatorResult.Success(endOfPaginationReached = postKey != null))
-
-                nextKey
-            }
-        }
-
-        return blogApi.getPosts(page).subscribeOn(Schedulers.io()).map { posts ->
-            appDatabase.runInTransaction {
-                if (loadType == LoadType.REFRESH) {
-                    postKeyDao.clearKeys()
-                    // clear posts
-                    Log.e("is Refresh", "About to clear posts table")
-                    postDao.deletePosts()
+                        remoteKeys.nextKey ?: -1
+                    }
                 }
+            }.flatMap { page ->
+                if(page == -1) {
+                    Single.just(MediatorResult.Success(endOfPaginationReached = true))
+                } else {
+                    blogApi.getPosts(page).subscribeOn(Schedulers.io()).map { posts ->
+                        appDatabase.runInTransaction {
+                            if (loadType == LoadType.REFRESH) {
+                                postKeyDao.clearKeys()
+                                // clear posts
+                                postDao.deletePosts()
+                            }
 
-                val prevKey = if (page == -1) null else page - 1
-                val nextKey = if (posts.isEmpty()) null else page + 1
+                            val prevKey = if (page == 1) null else page - 1
+                            val nextKey = if (posts.isEmpty()) null else page + 1
 
-                val keys = posts.map {
-                    PostKey(it.id.toLong(), prevKey, nextKey)
+                            val keys = posts.map {
+                                PostKey(it.id.toLong(), prevKey, nextKey)
+                            }
+
+
+                            postKeyDao.insertAll(keys)
+                            postDao.insertAll(posts)
+                        }
+
+                        return@map MediatorResult.Success(posts.isEmpty())
+                    }
                 }
-
-
-                postKeyDao.insertAll(keys)
-                postDao.insertAll(posts)
             }
-
-            return@map MediatorResult.Success(posts.isEmpty())
-        }
     }
 
     private fun getPostKeyForLastItem(state: PagingState<Int, Post>): PostKey? {
@@ -105,7 +102,7 @@ class PostsRemoteMediator(
         // Get the item closest to the anchor position
         return state.anchorPosition?.let { position ->
             state.closestItemToPosition(position)?.id?.let { postId ->
-                postKeyDao.postKeysById(postId.toLong())
+                return postKeyDao.postKeysById(postId.toLong())
             }
         }
     }
